Embed Command
This commit adds a command `embed` that enumerates recursively through a
directory and then generates a map where the key is the relative path name,
and the value is the byte array representing the file. Unlike other tools
that perform the same functionality, the output of this command introduces
no dependencies on other frameworks.

This implementation contains no filtering; every single file in the
specified directory and any subdirectories will be included in the map.